% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyjShiny.R
\name{addGraphFromDataFrame}
\alias{addGraphFromDataFrame}
\title{Add graph from data.frame}
\usage{
addGraphFromDataFrame(session, tbl.edges, tbl.nodes = NULL)
}
\arguments{
\item{session}{a Shiny Server session object.}

\item{tbl.edges}{a data.frame with source, traget, interaction columns (and option other attributes)}

\item{tbl.nodes}{(optional; nodes can be deduced from tbl.edges) a data.frame with nodes and their attributes}
}
\value{
Nothing
}
\description{
Add graph from data.frame
}
\examples{
\dontrun{
addGraphFromDataFrame(session)
}

}
