% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyjShiny.R
\name{addGraphFromJsonFile}
\alias{addGraphFromJsonFile}
\title{addGraphFromJsonFile}
\usage{
addGraphFromJsonFile(session, jsonFilename)
}
\arguments{
\item{session}{a Shiny Server session object.}

\item{jsonFilename}{of a text file with JSON representation of a cytoscape.js graph}
}
\description{
addGraphFromJsonFile
}
\examples{
\dontrun{
addGraphFromJsonFile(session)
}
}
