% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyjShiny.R
\name{clearSelection}
\alias{clearSelection}
\title{clearSelection all node and edge selections removed}
\usage{
clearSelection(session)
}
\arguments{
\item{session}{a Shiny Server session object.}
}
\description{
clearSelection all node and edge selections removed
}
