% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createCytoscapeJsNetwork.R
\name{createCytoscapeJsNetwork}
\alias{createCytoscapeJsNetwork}
\title{Generate a CytoscapeJS compatible network}
\usage{
createCytoscapeJsNetwork(nodeData, edgeData, nodeColor = "#888888",
  nodeShape = "ellipse", nodeHeight = "70", nodeWidth = "70",
  nodeLabelColor = "#FFFFFF", edgeColor = "#888888",
  edgeSourceShape = "none", edgeTargetShape = "triangle", nodeHref = "")
}
\arguments{
\item{nodeData}{a data.frame with at least two columns: id and name; Any additional 
columns from nodeData will be passed to CytoscapeJS. If 'x' and 'y' columns are 
specified and the layout in rcytoscapejs function is set 
to 'preset' then the user-specified locations will be used.}

\item{edgeData}{a data.frame with at least two columns: source and target}

\item{nodeColor}{a hex color for nodes (default: #666666)}

\item{nodeShape}{a shape for nodes (default: ellipse)}

\item{nodeHeight}{node height (default: 30)}

\item{nodeWidth}{node width (default: 30)}

\item{nodeLabelColor}{node label color (default: #FFFFFF)}

\item{edgeColor}{a hex color for edges (default: #666666)}

\item{edgeSourceShape}{a shape for arrow sources (default: none)}

\item{edgeTargetShape}{a shape for arrow targets (default: triangle)}

\item{nodeHref}{a link associated with the node (appears in a tooltip)}
}
\value{
a list with two entries:

\itemize{
  \item nodes a JSON string with node information compatible with CytoscapeJS
  \item edges a JSON string with edge information compatible with CytoscapeJS
}

If required columns (see parameters) are not provided, an error will be thrown

See http://cytoscape.github.io/cytoscape.js/ for shape details
}
\description{
Generate a CytoscapeJS compatible network
}
\examples{
id <- c("Jerry", "Elaine", "Kramer", "George")
name <- id
nodeData <- data.frame(id, name, stringsAsFactors=FALSE)

source <- c("Jerry", "Jerry", "Jerry", "Elaine", "Elaine", "Kramer", "Kramer", "Kramer", "George")
target <- c("Elaine", "Kramer", "George", "Jerry", "Kramer", "Jerry", "Elaine", "George", "Jerry")
edgeData <- data.frame(source, target, stringsAsFactors=FALSE)

network <- createCytoscapeJsNetwork(nodeData, edgeData)

}
\seealso{
\code{\link{rcytoscapejs}}
}

