% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphsToJSON.R
\name{dataFramesToJSON}
\alias{dataFramesToJSON}
\title{Create a cytoscape.js JSON graph from one or two data.frames.}
\usage{
dataFramesToJSON(tbl.edges, tbl.nodes = NULL)
}
\arguments{
\item{tbl.edges}{data.frame, with source, target and interaction columns, others option for edge attributes}

\item{tbl.nodes}{data.frame, options, useful for orphan nodes, and necessary for adding node attributes}
}
\description{
Create a cytoscape.js JSON graph from one or two data.frames.
}
