% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyjShiny.R
\name{doLayout}
\alias{doLayout}
\title{Layout the current graph using the specified strategy.}
\usage{
doLayout(session, strategy)
}
\arguments{
\item{session}{a Shiny Server session object}

\item{layoutStrategy}{character string, one of cola, cose, circle, concentric, grid, breadthfirst, random, dagre, cose-bilkent}
}
\description{
Layout the current graph using the specified strategy.
}
\examples{
  doLayout(session, "cola")

}
