% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyjShiny.R
\name{fit}
\alias{fit}
\title{Set zoom and center of the graph display so that graph fills the display}
\usage{
fit(session)
}
\arguments{
\item{session}{a Shiny server session object}
}
\description{
Set zoom and center of the graph display so that graph fills the display
}
\examples{

  fit(session)

}
