% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyjShiny.R
\name{getNodePositions}
\alias{getNodePositions}
\title{Get node positions}
\usage{
getNodePositions(session)
}
\arguments{
\item{session}{a Shiny Server session object.}
}
\value{
Nothing
}
\description{
Get node positions
}
