% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyjShiny.R
\name{getSelectedNodes}
\alias{getSelectedNodes}
\title{getSelectedNodes}
\usage{
getSelectedNodes(session)
}
\arguments{
\item{session}{a Shiny server session object.}
}
\value{
a data.frame with (at least) an id column
}
\description{
\code{getSelectedNodes} get the selected nodes
}
