% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphsToJSON.R
\name{graphNELtoJSON}
\alias{graphNELtoJSON}
\title{Convert R graphNEL object to cytoscape.js JSON.}
\usage{
graphNELtoJSON(g)
}
\arguments{
\item{g}{a graphNEL}
}
\value{
a string with a cytoscape.js JSON graph
}
\description{
Convert R graphNEL object to cytoscape.js JSON.
}
\examples{
\dontrun{
g.json <- graphNELtoJSON(graphNEL())
}

}
