% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyjShiny.R
\name{graphToJSON}
\alias{graphToJSON}
\title{Convert R graphNEL object to cytoscape.js JSON.}
\usage{
graphToJSON(g)
}
\arguments{
\item{g}{a graphNEL}
}
\description{
Convert R graphNEL object to cytoscape.js JSON.
}
\examples{
\dontrun{
  g.json <- graphToJSON(graphNEL())
}

}
