% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyjShiny.R
\name{hideSelection}
\alias{hideSelection}
\title{Hide selection all selected nodes and their edges are hidden}
\usage{
hideSelection(session)
}
\arguments{
\item{session}{a Shiny Server session object.}
}
\value{
Nothing
}
\description{
Hide selection all selected nodes and their edges are hidden
}
