% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcytoscapejs.R
\name{rcytoscapejs}
\alias{rcytoscapejs}
\title{Create CytoscapeJS widget}
\usage{
rcytoscapejs(nodeEntries, edgeEntries, layout = "cose", width = NULL,
  height = NULL, showPanzoom = TRUE, highlightConnectedNodes = TRUE,
  boxSelectionEnabled = TRUE)
}
\arguments{
\item{nodeEntries}{nodes output from createCytoscapeJsNetwork function}

\item{edgeEntries}{edges output from createCytoscapeJsNetwork function}

\item{layout}{the layout algorithm to use for layout. If 'preset', then columns
'x' and 'y' from nodeEntries will be used for position information.}

\item{width}{width of widget (Default: 100 percent for Shiny apps)}

\item{height}{height of widget (Default: 400px for Shiny apps)}

\item{showPanzoom}{whether to show the CytoscapeJS panzoom menu in the upper left corner (Default: TRUE)}

\item{highlightConnectedNodes}{whether to highlight connected nodes to a clicked node (Default: TRUE)}

\item{boxSelectionEnabled}{whether to allow users to box select multiple nodes (Default: TRUE)}
}
\description{
Pass network information to CytoscapeJS widget
}
\examples{
id <- c("Jerry", "Elaine", "Kramer", "George")
name <- id
nodeData <- data.frame(id, name, stringsAsFactors=FALSE)

source <- c("Jerry", "Jerry", "Jerry", "Elaine", "Elaine", "Kramer", "Kramer", "Kramer", "George")
target <- c("Elaine", "Kramer", "George", "Jerry", "Kramer", "Jerry", "Elaine", "George", "Jerry")
edgeData <- data.frame(source, target, stringsAsFactors=FALSE)

network <- createCytoscapeJsNetwork(nodeData, edgeData)
rcytoscapejs(network$nodes, network$edges)

}
\seealso{
\code{\link{createCytoscapeJsNetwork}}
}

