% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcytoscapejs.R
\name{rcytoscapejsOutput}
\alias{rcytoscapejsOutput}
\title{Widget output function for use in Shiny}
\usage{
rcytoscapejsOutput(outputId, width = "100\%", height = "400px")
}
\arguments{
\item{outputId}{a string identifying the Shiny output}

\item{width}{width of the widget}

\item{height}{height of the widget}
}
\description{
Widget output function for use in Shiny
}

