% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphsToJSON.R
\name{readAndStandardizeJSONNetworkFile}
\alias{readAndStandardizeJSONNetworkFile}
\title{Read in a JSON network file, identify (or add) elements field return JSON}
\usage{
readAndStandardizeJSONNetworkFile(filename)
}
\arguments{
\item{file}{a json file}
}
\description{
Read in a JSON network file, identify (or add) elements field return JSON
}
