% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyjShiny.R
\name{removeGraph}
\alias{removeGraph}
\title{remove the current graph}
\usage{
removeGraph(session)
}
\arguments{
\item{session}{a Shiny Server session object.}
}
\description{
remove the current graph
}
\examples{
\dontrun{
  removeGraph(session)
}
}
