% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyjShiny.R
\name{renderCyjShiny}
\alias{renderCyjShiny}
\title{More shiny plumbing - now sure exactly what this does}
\usage{
renderCyjShiny(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{expr}{not sure...}

\item{env}{environment not sure ...}

\item{quoteed}{logical not sure ...}
}
\value{
not sure
}
\description{
More shiny plumbing - now sure exactly what this does
}
