% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcytoscapejs.R
\name{renderRcytoscapejs}
\alias{renderRcytoscapejs}
\title{Widget render function for use in Shiny}
\usage{
renderRcytoscapejs(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{expr}{An expression that returns an R object that can be used as an argument to cat.}

\item{env}{The environment in which to evaluate expr.}

\item{quoted}{Is expr a quoted expression (with quote())? This is useful if you want to save an expression in a variable.}
}
\description{
Widget render function for use in Shiny
}

