% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runShinyApp.R
\name{runShinyApp}
\alias{runShinyApp}
\title{Run Shiny App}
\usage{
runShinyApp(launch.browser = TRUE, port = 3838)
}
\arguments{
\item{launch.browser}{launch in browser? (default: TRUE)}

\item{port}{port number to use}
}
\value{
None
}
\description{
Run Shiny App
}
\examples{
port <-3838
# Uncomment first
#runShinyApp(port=port)

}

