% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyjShiny.R
\name{savePNGtoFile}
\alias{savePNGtoFile}
\title{save a png rendering of the current network view to the specified filename}
\usage{
savePNGtoFile(session, filename)
}
\arguments{
\item{session}{a Shiny Server session object.}

\item{filename}{a character string}
}
\description{
save a png rendering of the current network view to the specified filename
}
