% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyjShiny.R
\name{selectFirstNeighbors}
\alias{selectFirstNeighbors}
\title{selectFirstNeighbors of the currently selected nodes}
\usage{
selectFirstNeighbors(session)
}
\arguments{
\item{session}{a Shiny Server session object.}
}
\description{
selectFirstNeighbors of the currently selected nodes
}
