% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyjShiny.R
\name{selectNodes}
\alias{selectNodes}
\title{Select Nodes}
\usage{
selectNodes(session, nodeNames)
}
\arguments{
\item{session}{a Shiny Server session object.}

\item{nodeNames}{character, a list of node IDs}
}
\value{
Nothing
}
\description{
Select Nodes
}
