% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyjShiny.R
\name{setNodePositions}
\alias{setNodePositions}
\title{Set node positions from the supplied data.frame}
\usage{
setNodePositions(session, tbl.positions)
}
\arguments{
\item{session}{a Shiny Server session object.}

\item{tbl.positions}{a data.frame with three columns: id, x, y}
}
\value{
Nothing
}
\description{
Set node positions from the supplied data.frame
}
