% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyjShiny.R
\name{addGraphFromJsonFile}
\alias{addGraphFromJsonFile}
\title{Add graph from JSON file}
\usage{
addGraphFromJsonFile(session, jsonFilename)
}
\arguments{
\item{session}{a Shiny Server session object.}

\item{jsonFilename}{of a text file with JSON representation of a cytoscape.js graph}
}
\value{
Nothing
}
\description{
Add graph from JSON file
}
\examples{
\dontrun{
addGraphFromJsonFile(session)
}
}
