% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyjShiny.R
\name{cyjShinyOutput}
\alias{cyjShinyOutput}
\title{Standard shiny ui rendering construct}
\usage{
cyjShinyOutput(outputId, width = "100\%", height = "400")
}
\arguments{
\item{outputId}{the name of the DOM element to create.}

\item{width}{integer  optional initial width of the widget.}

\item{height}{integer optional initial height of the widget.}
}
\value{
a reference to an htmlwidget
}
\description{
Standard shiny ui rendering construct
}
\examples{
\dontrun{
mainPanel(cyjShinyOutput("cyjShiny"), width = 10)
}

}
