% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyjShiny.R
\name{fit}
\alias{fit}
\title{Set zoom and center of the graph display so that graph fills the display.}
\usage{
fit(session, padding = 50)
}
\arguments{
\item{session}{a Shiny server session object.}

\item{padding}{integer, default 50 pixels.}
}
\value{
Nothing
}
\description{
Set zoom and center of the graph display so that graph fills the display.
}
\examples{
\dontrun{
fit(session, 100)
}

}
\seealso{
\code{\link{fitSelected}}
}
