% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyjShiny.R
\name{fitSelected}
\alias{fitSelected}
\title{Set zoom and center of the graph display so that the currently selected nodes fill the display}
\usage{
fitSelected(session, padding = 50)
}
\arguments{
\item{session}{a Shiny server session object.}

\item{padding}{integer, default 50 pixels.}
}
\value{
Nothing
}
\description{
Set zoom and center of the graph display so that the currently selected nodes fill the display
}
\examples{
\dontrun{
fitSelected(session, 100)
}

}
\seealso{
\code{\link{fit}}
}
