% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyjShiny.R
\name{invertSelection}
\alias{invertSelection}
\title{Invert selection all selected nodes and their edges are hidden}
\usage{
invertSelection(session)
}
\arguments{
\item{session}{a Shiny Server session object.}
}
\value{
Nothing
}
\description{
Invert selection all selected nodes and their edges are hidden
}
