% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyjShiny.R
\name{loadNetworkFromJSONFile}
\alias{loadNetworkFromJSONFile}
\title{Load a standard cytoscape.js JSON network file}
\usage{
loadNetworkFromJSONFile(filename)
}
\arguments{
\item{filename}{character string, either relative or absolute path.}
}
\value{
Nothing
}
\description{
Load a standard cytoscape.js JSON network file
}
\examples{
\dontrun{
loadNetworkFromJSONFile(system.file(package = "cyjShiny", "extdata", "galFiltered.cyjs"))
}

}
