% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyjShiny.R
\name{loadStyleFile}
\alias{loadStyleFile}
\title{Load a standard cytoscape.js style file}
\usage{
loadStyleFile(styleFile)
}
\arguments{
\item{styleFile}{character string, either relative or absolute path.}
}
\value{
Nothing
}
\description{
Load a standard cytoscape.js style file
}
\examples{
\dontrun{
loadStyleFile(system.file(package = "cyjShiny", "extdata", "yeastGalactoseStyle.js"))
}

}
