% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphsToJSON.R
\name{readAndStandardizeJSONStyleFile}
\alias{readAndStandardizeJSONStyleFile}
\title{Read in a JSON file, extract the selector elements, return JSON}
\usage{
readAndStandardizeJSONStyleFile(filename)
}
\arguments{
\item{file}{a json file}
}
\description{
Read in a JSON file, extract the selector elements, return JSON
}
